adb restructured
