+++ src/exojax/database/core_atom/broadening.py
@@ -4,7 +4,7 @@ import jax.numpy as jnp
 import numpy as np
 from jax.lax import scan
 
-from exojax.database import atomllapi
+from exojax.database import core_atom.io
 from exojax.utils.constants import Rcgs
 from exojax.utils.constants import a0
 from exojax.utils.constants import ccgs
